Added redirect to error page and included to tests changes in accordance with the commentary to the PR
