+++ apps/site/src/app/components/index.ts
@@ -1,2 +1,2 @@
+export * from "../../utils/routing";
 export * from ".";
-export * from "./routing";
